% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMoveData.R
\name{postPredMovements}
\alias{postPredMovements}
\title{A function to plot posterior predicitve checks on the model}
\usage{
postPredMovements(model.fit, df.obs, rangePred = 1:10, numbDraws = 500,
  buffer = 10)
}
\arguments{
\item{model.fit}{stanfit model, containing y_pred and d_pred as generated quantities: y_pred is the model predicted angle, and d_pred is the model predicted step length.}

\item{df.obs}{dataframe of observed travel points.}

\item{rangePred}{Range of observed travel points to predict.}

\item{numbDraws}{Number of predictions to make for each observed point.}

\item{buffer}{Spatial buffer for the extent of the output plot around observed points.}
}
\description{
This function uses samples from the stanfit object to generate predictive regions in 2d.
}
