% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rquaredAngles.R
\name{r.squared.angles}
\alias{r.squared.angles}
\title{A function to estimate explained variance}
\usage{
r.squared.angles(model.fit, obs.angles)
}
\arguments{
\item{model.fit}{stanfit model, containing a_pred as generated quantities: a_pred is the model predicted angle.}

\item{obs.angles}{Vector of observed angles used in model fitting.}
}
\description{
This function uses samples from a stanfit object to estimate the error between predicted and observed angles. R-squared is then estimated as 1 minus the variation in the error divided by the total variation in the observed angles.
}
\examples{


}
