% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMoveData.R
\name{simData.plot}
\alias{simData.plot}
\title{A function to plot the simulated data}
\usage{
simData.plot(simData, max.tracks = 5)
}
\arguments{
\item{simData}{data generated by simulateData.}

\item{max.tracks}{limits the plots to tracks below this number.}
}
\description{
This function is just a convient way to plot the simulated data.
}
\examples{
sim.df <- simulateData()
simData.plot(sim.df)


}
