% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moveModel.R
\name{simMovements}
\alias{simMovements}
\title{A function to fit movement models using rStan}
\usage{
simMovements(modelList, iter = 100, startingContext = c(0, 0),
  names = c("x", "y"), formula = "a ~ beta_cv * cv", functionList = NULL)
}
\arguments{
\item{modelList}{A list of stanfit models.}
}
\description{
This function fits standard movement models using rstan to observed movement data. STOPPED HERE
}
