% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMoveData.R
\name{simulateData}
\alias{simulateData}
\title{A function to simulate movement data}
\usage{
simulateData(nobs = 1000, ntracks = 10, meanStepLength = 2,
  sdStepLength = 0.25, influenceRange = 50, EffectStrength = 0.5,
  bearingSD = 5)
}
\arguments{
\item{nobs}{number of observations. Defaults to 1000.}

\item{ntracks}{number of observed movement tracks. Defaults to 10.}

\item{meanStepLength}{mean of step lengths from a lognormal distribution. Defaults to 2.}

\item{sdStepLength}{sd of step lengths from a lognormal distribution. Defaults to 0.25.}

\item{influenceRange}{range in which influence point is randomly assigned. Defaults to 50.}

\item{EffectStrength}{strength of influence relative to directional persistance. Defaults to 0.5.}

\item{bearingSD}{magnitude of error in directional persistance. Defaults to 5.}
}
\description{
This function allows you to simulate movement data, where an individual animal is influenced by an external influence. The strength of this influence can be set by the user.
}
\examples{
simulateData()


}
\keyword{movement}
\keyword{simulation,}
