% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMoveData.R
\name{postPredAngle}
\alias{postPredAngle}
\title{A function to plot posterior predicitve checks on directions of travel}
\usage{
postPredAngle(model.fit, df.obs, contours = c(25, 50, 75), rangePred = 1:10,
  numbDraws = 500, buffer = 10, extention = 10, context = NULL)
}
\arguments{
\item{model.fit}{stanfit model, containing a_pred as generated quantities: a_pred is the model predicted angle.}

\item{df.obs}{dataframe of observed travel points.}

\item{contours}{Contours used in the plot of the kernel density of predicted points, if null a raster is produced.}

\item{rangePred}{Range of observed travel points to predict.}

\item{numbDraws}{Number of predictions to make for each observed point.}

\item{buffer}{Display parameter, used to extend the plot extent range.}

\item{extention}{Display parameter, used to define the radius on which to plot predictions for each point.}

\item{context}{List of shapefiles or rasters that can provide contextual information.}
}
\description{
This function uses samples from the stanfit object to generate predictive directions.
}
